Refactor: Sidebar 컴포넌트에서 로직 custom hook 분리
