+++ src/redux/slice/editCourseIdSlice.ts
@@ -52,7 +52,7 @@ const editCourseSlice = createSlice({
     setLectureCount: (state, action: PayloadAction<number>) => {
       state.lectureCount = action.payload;
     },
- 
+
     // 선택된 몇 번째 course가 선택된건지 체크
     setSelectedCourse: (state, action) => {
       state.selectedCourse = action.payload;
